Migrate Android XR APIs from Godot core to the Vendors plugin
