+++ plugin/src/main/cpp/export/khronos_export_plugin.cpp
@@ -103,22 +103,6 @@ Dictionary KhronosEditorExportPlugin::_get_export_options_overrides(const Ref<go
 	return overrides;
 }
 
-PackedStringArray KhronosEditorExportPlugin::_get_export_features(const Ref<EditorExportPlatform> &platform, bool debug) const {
-	PackedStringArray features;
-	if (!_supports_platform(platform) || !_is_vendor_plugin_enabled()) {
-		return features;
-	}
-
-	if (_is_khronos_htc_enabled()) {
-		// Add the eye tracking feature if necessary
-		if ((bool)ProjectSettings::get_singleton()->get_setting_with_override("xr/openxr/extensions/eye_gaze_interaction")) {
-			features.append(EYE_GAZE_INTERACTION_FEATURE);
-		}
-	}
-
-	return features;
-}
-
 String KhronosEditorExportPlugin::_get_export_option_warning(const Ref<EditorExportPlatform> &platform, const String &option) const {
 	if (!_supports_platform(platform) || !_is_vendor_plugin_enabled()) {
 		return "";
