+++ plugin/src/main/cpp/export/meta_export_plugin.cpp
@@ -228,14 +228,9 @@ PackedStringArray MetaEditorExportPlugin::_get_supported_devices() const {
 }
 
 PackedStringArray MetaEditorExportPlugin::_get_export_features(const Ref<EditorExportPlatform> &platform, bool debug) const {
-	PackedStringArray features;
+	PackedStringArray features = OpenXRVendorsEditorExportPlugin::_get_export_features(platform, debug);
 	if (!_supports_platform(platform) || !_is_vendor_plugin_enabled()) {
 		return features;
-	}
-
-	// Add the eye tracking feature if necessary
-	if ((bool)ProjectSettings::get_singleton()->get_setting_with_override("xr/openxr/extensions/eye_gaze_interaction")) {
-		features.append(EYE_GAZE_INTERACTION_FEATURE);
 	}
 
 	// Add a feature to indicate that this is a hybrid app.
