+++ plugin/src/main/cpp/export/pico_export_plugin.cpp
@@ -74,20 +74,6 @@ TypedArray<Dictionary> PicoEditorExportPlugin::_get_export_options(const Ref<Edi
 	return export_options;
 }
 
-PackedStringArray PicoEditorExportPlugin::_get_export_features(const Ref<EditorExportPlatform> &platform, bool debug) const {
-	PackedStringArray features;
-	if (!_supports_platform(platform) || !_is_vendor_plugin_enabled()) {
-		return features;
-	}
-
-	// Add the eye tracking feature if necessary
-	if ((bool)ProjectSettings::get_singleton()->get_setting_with_override("xr/openxr/extensions/eye_gaze_interaction")) {
-		features.append(EYE_GAZE_INTERACTION_FEATURE);
-	}
-
-	return features;
-}
-
 String PicoEditorExportPlugin::_get_export_option_warning(const Ref<EditorExportPlatform> &platform, const String &option) const {
 	if (!_supports_platform(platform) || !_is_vendor_plugin_enabled()) {
 		return "";
